// frontend/src/generated/MessageFilters.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DateRange } from './DateRange'

export type MessageFilters = {
  range: DateRange
  timezone: string
  hostId: string | null
  source: string | null
  agentKey: string | null
  providerId: string | null
  modelId: string | null
  isIncomplete: boolean | null
}